requirements.txt
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DAWG2-Python==0.9.0
feedparser==6.0.12
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.16.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydeck==0.9.1
pymorphy3==2.0.6
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wrapt==2.1.1
